+++ pkg-r/R/chat_restore.R
@@ -92,8 +92,9 @@ chat_restore <- function(
   # Exclude works with bookmark names
   excluded_names <- session$getBookmarkExclude()
   id_user_input <- paste0(id, "_user_input")
-  if (!(id_user_input %in% excluded_names)) {
-    session$setBookmarkExclude(c(excluded_names, id_user_input))
+  to_exclude <- setdiff(id_user_input, excluded_names)
+  if (length(to_exclude) > 0) {
+    session$setBookmarkExclude(c(excluded_names, to_exclude))
   }
 
   # Save